+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - No city filter
 - Search query filter
 - Zoom to city on selection
+- Add link to Google Maps search for playground name
 
 ### Changed
 - Improve city determination